+++ src/NzbDrone.Integration.Test/ApiTests/WantedTests/CutoffUnmetFixture.cs
@@ -5,10 +5,10 @@ using NzbDrone.Core.Books;
 using NzbDrone.Core.Qualities;
 using Readarr.Api.V1.RootFolders;
 
-namespace NzbDrone.Integration.Test.ApiTests
+namespace NzbDrone.Integration.Test.ApiTests.WantedTests
 {
     [TestFixture]
-    public class WantedFixture : IntegrationTest
+    public class CutoffUnmetFixture : IntegrationTest
     {
         [SetUp]
         public void Setup()
@@ -25,55 +25,10 @@ namespace NzbDrone.Integration.Test.ApiTests
         }
 
         [Test]
-        [Order(0)]
-        public void missing_should_be_empty()
-        {
-            EnsureNoAuthor("14586394", "Andrew Hunter Murray");
-
-            var result = WantedMissing.GetPaged(0, 15, "releaseDate", "desc");
-
-            result.Records.Should().BeEmpty();
-        }
-
-        [Test]
-        [Order(1)]
-        public void missing_should_have_monitored_items()
-        {
-            EnsureAuthor("14586394", "43765115", "Andrew Hunter Murray", true);
-
-            var result = WantedMissing.GetPaged(0, 15, "releaseDate", "desc");
-
-            result.Records.Should().NotBeEmpty();
-        }
-
-        [Test]
-        [Order(1)]
-        public void missing_should_have_author()
-        {
-            EnsureAuthor("14586394", "43765115", "Andrew Hunter Murray", true);
-
-            var result = WantedMissing.GetPagedIncludeAuthor(0, 15, "releaseDate", "desc", includeAuthor: true);
-
-            result.Records.First().Author.Should().NotBeNull();
-            result.Records.First().Author.AuthorName.Should().Be("Andrew Hunter Murray");
-        }
-
-        [Test]
-        [Order(1)]
-        public void missing_should_not_have_author()
-        {
-            EnsureAuthor("14586394", "43765115", "Andrew Hunter Murray", true);
-
-            var result = WantedMissing.GetPagedIncludeAuthor(0, 15, "releaseDate", "desc", includeAuthor: false);
-
-            result.Records.First().Author.Should().BeNull();
-        }
-
-        [Test]
         [Order(2)]
         public void cutoff_should_have_monitored_items()
         {
-            EnsureProfileCutoff(1, Quality.AZW3);
+            EnsureProfileCutoff(1, Quality.AZW3, true);
             var author = EnsureAuthor("14586394", "43765115", "Andrew Hunter Murray", true);
             EnsureBookFile(author, 1, "43765115", Quality.MOBI);
 
@@ -83,21 +38,10 @@ namespace NzbDrone.Integration.Test.ApiTests
         }
 
         [Test]
-        [Order(1)]
-        public void missing_should_not_have_unmonitored_items()
-        {
-            EnsureAuthor("14586394", "43765115", "Andrew Hunter Murray", false);
-
-            var result = WantedMissing.GetPaged(0, 15, "releaseDate", "desc");
-
-            result.Records.Should().BeEmpty();
-        }
-
-        [Test]
         [Order(2)]
         public void cutoff_should_not_have_unmonitored_items()
         {
-            EnsureProfileCutoff(1, Quality.AZW3);
+            EnsureProfileCutoff(1, Quality.AZW3, true);
             var author = EnsureAuthor("14586394", "43765115", "Andrew Hunter Murray", false);
             EnsureBookFile(author, 1, "43765115", Quality.MOBI);
 
@@ -110,7 +54,7 @@ namespace NzbDrone.Integration.Test.ApiTests
         [Order(2)]
         public void cutoff_should_have_author()
         {
-            EnsureProfileCutoff(1, Quality.AZW3);
+            EnsureProfileCutoff(1, Quality.AZW3, true);
             var author = EnsureAuthor("14586394", "43765115", "Andrew Hunter Murray", true);
             EnsureBookFile(author, 1, "43765115", Quality.MOBI);
 
@@ -124,7 +68,7 @@ namespace NzbDrone.Integration.Test.ApiTests
         [Order(2)]
         public void cutoff_should_not_have_author()
         {
-            EnsureProfileCutoff(1, Quality.AZW3);
+            EnsureProfileCutoff(1, Quality.AZW3, true);
             var author = EnsureAuthor("14586394", "43765115", "Andrew Hunter Murray", true);
             EnsureBookFile(author, 1, "43765115", Quality.MOBI);
 
@@ -134,21 +78,10 @@ namespace NzbDrone.Integration.Test.ApiTests
         }
 
         [Test]
-        [Order(1)]
-        public void missing_should_have_unmonitored_items()
-        {
-            EnsureAuthor("14586394", "43765115", "Andrew Hunter Murray", false);
-
-            var result = WantedMissing.GetPaged(0, 15, "releaseDate", "desc", "monitored", "false");
-
-            result.Records.Should().NotBeEmpty();
-        }
-
-        [Test]
         [Order(2)]
         public void cutoff_should_have_unmonitored_items()
         {
-            EnsureProfileCutoff(1, Quality.AZW3);
+            EnsureProfileCutoff(1, Quality.AZW3, true);
             var author = EnsureAuthor("14586394", "43765115", "Andrew Hunter Murray", false);
             EnsureBookFile(author, 1, "43765115", Quality.MOBI);
